Added JoinSqlFullAddressService and integrated with DatabaseController
Added a new dependency `JoinSqlFullAddressService` to the `DatabaseController` class and initialized it in the constructor. A new HTTP POST method `JoinAsync` has been added to the `DatabaseController` class, which calls the `JoinFullAddressAsync` method of the `JoinSqlFullAddressService` class. The `JoinSqlFullAddressService` has been added as a singleton service in the application's dependency injection container in the `Program.cs` file. A new class `JoinSqlFullAddressService` has been created, which has a method `JoinFullAddressAsync` that executes a SQL command to create a new table `FullAddress` if it does not exist. This table is created by joining data from multiple tables and aggregating the full address. The execution time of this operation is logged.